Scale down site-wide og:image & add light BG.
Bluesky can't process Open Graph images larger than 1MB, and for the
moment, this actually throws an error that prevents the user from
a link with that image in its head. This is bizarre, but unclear if it
will change or if other platforms have similarly low thresholds.

As the OP here states, Mastodon supports up to 16MB, equally bizarre:
https://github.com/bluesky-social/atproto/discussions/1740

See also this, regarding the UX for this bug:
https://github.com/bluesky-social/atproto/discussions/1840

And here's the culprit:
https://github.com/bluesky-social/atproto/blob/a5b20f0218bd13e3c5d7681de2263dcc850b7523/lexicons/app/bsky/embed/images.json#L21-L25

I've also added a light background color, instead of transparency, so
the og:image displays more consistently across platforms.